requirements.txt
Flask==3.0.3
googletrans==4.0.0-rc1
Pillow==11.0.0
pytesseract==0.3.13
